agent/requirements.txt
google-genai==1.52.0
httpx>=0.28.1
redis==5.0.1
psycopg[binary]==3.1.18


